# backend/src/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Graphfolio Backend API
  description: Backend API for Graphfolio - Company network visualization
  version: 1.0.0
servers:
- url: https://api.tinboker.com
  description: Production server
- url: https://staging-api.tinboker.com
  description: Staging server
- url: https://dev-api.tinboker.com
  description: Development server
- url: http://localhost:8000
  description: Local development
paths:
...
  /api/admin/auth/login:
    post:
      tags:
      - admin
      summary: Admin Login
      description: 'Authenticate admin user with password.

        Returns JWT token on success.'
      operationId: admin_login_api_admin_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/translations/stats:
    get:
      tags:
      - admin
      summary: Get Translation Stats
      description: Get translation statistics.
      operationId: get_translation_stats_api_admin_translations_stats_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /api/admin/translations/missing:
    get:
      tags:
      - admin
      summary: Get Missing Translations
      description: Get translations without ZH-TW name.
      operationId: get_missing_translations_api_admin_translations_missing_get
      security:
      - HTTPBearer: []
      parameters:
      - name: market
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by market
          title: Market
        description: Filter by market
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          description: Max items to return
          default: 100
          title: Limit
        description: Max items to return
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissingTranslationsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/translations:
    get:
      tags:
      - admin
      summary: List Translations
      description: List translations with optional filters and pagination.
      operationId: list_translations_api_admin_translations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: market
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by market
          title: Market
        description: Filter by market
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by translation status
          title: Status
        description: Filter by translation status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search in ticker/name
          title: Search
        description: Search in ticker/name
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number
          default: 1
          title: Page
        description: Page number
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Items per page
          default: 50
          title: Limit
        description: Items per page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - admin
      summary: Create Translation
      description: Create a new translation.
      operationId: create_translation_api_admin_translations_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/translations/{translation_id}:
    get:
      tags:
      - admin
      summary: Get Translation
      description: Get a single translation by ID.
      operationId: get_translation_api_admin_translations__translation_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: translation_id
        in: path
        required: true
        schema:
          type: integer
          title: Translation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - admin
      summary: Update Translation
      description: Update an existing translation.
      operationId: update_translation_api_admin_translations__translation_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: translation_id
        in: path
        required: true
        schema:
          type: integer
          title: Translation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslationUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      summary: Delete Translation
      description: Delete a translation.
      operationId: delete_translation_api_admin_translations__translation_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: translation_id
        in: path
        required: true
        schema:
          type: integer
          title: Translation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/admin/translations/bulk-import:
    post:
      tags:
      - admin
      summary: Bulk Import Translations
      description: 'Bulk import translations from CSV file.

        Expected columns: ticker, market, name_en, name_zh_tw

        Optional column: translation_status (defaults to ''auto'')'
      operationId: bulk_import_translations_api_admin_translations_bulk_import_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_bulk_import_translations_api_admin_translations_bulk_import_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkImportResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/admin/translations/bulk-json:
    post:
      tags:
      - admin
      summary: Bulk Import Json
      description: Bulk import translations from JSON array.
      operationId: bulk_import_json_api_admin_translations_bulk_json_post
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/BulkImportItem'
              type: array
              title: Items
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkImportResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    Body_bulk_import_translations_api_admin_translations_bulk_import_post:
      properties:
        file:
          type: string
          format: binary
          title: File
          description: CSV file with translations
      type: object
      required:
      - file
      title: Body_bulk_import_translations_api_admin_translations_bulk_import_post
...
    BulkImportItem:
      properties:
        ticker:
          type: string
          maxLength: 20
          title: Ticker
          description: Stock ticker symbol
        market:
          type: string
          maxLength: 10
          title: Market
          description: Market code (US, TW, JP)
        name_en:
          anyOf:
          - type: string
          - type: 'null'
          title: Name En
          description: English name
        name_zh_tw:
          anyOf:
          - type: string
          - type: 'null'
          title: Name Zh Tw
          description: Chinese Traditional name
        translation_status:
          type: string
          enum:
          - pending
          - approved
          - auto
          title: Translation Status
          default: auto
      type: object
      required:
      - ticker
      - market
      title: BulkImportItem
      description: Schema for a single item in bulk import.
    BulkImportResponse:
      properties:
        imported:
          type: integer
          title: Imported
        updated:
          type: integer
          title: Updated
        errors:
          items:
            type: string
          type: array
          title: Errors
          default: []
      type: object
      required:
      - imported
      - updated
      title: BulkImportResponse
      description: Schema for bulk import response.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        password:
          type: string
          title: Password
      type: object
      required:
      - password
      title: LoginRequest
      description: Request schema for admin login.
    LoginResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in:
          type: integer
          title: Expires In
      type: object
      required:
      - access_token
      - expires_in
      title: LoginResponse
      description: Response schema for admin login.
...
    MissingTranslation:
      properties:
        ticker:
          type: string
          title: Ticker
        market:
          type: string
          title: Market
        name_en:
          anyOf:
          - type: string
          - type: 'null'
          title: Name En
      type: object
      required:
      - ticker
      - market
      title: MissingTranslation
      description: Schema for missing translation report.
    MissingTranslationsResponse:
      properties:
        total:
          type: integer
          title: Total
        items:
          items:
            $ref: '#/components/schemas/MissingTranslation'
          type: array
          title: Items
      type: object
      required:
      - total
      - items
      title: MissingTranslationsResponse
      description: Schema for missing translations list.
...
    TranslationCreate:
      properties:
        ticker:
          type: string
          maxLength: 20
          title: Ticker
          description: Stock ticker symbol
        market:
          type: string
          maxLength: 10
          title: Market
          description: Market code (US, TW, JP)
        name_en:
          anyOf:
          - type: string
          - type: 'null'
          title: Name En
          description: English name
        name_zh_tw:
          anyOf:
          - type: string
          - type: 'null'
          title: Name Zh Tw
          description: Chinese Traditional name
        translation_status:
          type: string
          enum:
          - pending
          - approved
          - auto
          title: Translation Status
          default: pending
      type: object
      required:
      - ticker
      - market
      title: TranslationCreate
      description: Schema for creating a new translation.
    TranslationListResponse:
      properties:
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        limit:
          type: integer
          title: Limit
        items:
          items:
            $ref: '#/components/schemas/TranslationResponse'
          type: array
          title: Items
      type: object
      required:
      - total
      - page
      - limit
      - items
      title: TranslationListResponse
      description: Schema for paginated translation list.
...
    TranslationResponse:
      properties:
        ticker:
          type: string
          maxLength: 20
          title: Ticker
          description: Stock ticker symbol
        market:
          type: string
          maxLength: 10
          title: Market
          description: Market code (US, TW, JP)
        name_en:
          anyOf:
          - type: string
          - type: 'null'
          title: Name En
          description: English name
        name_zh_tw:
          anyOf:
          - type: string
          - type: 'null'
          title: Name Zh Tw
          description: Chinese Traditional name
        id:
          type: integer
          title: Id
        translation_status:
          type: string
          title: Translation Status
        last_updated_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Updated By
        last_updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Updated At
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
      type: object
      required:
      - ticker
      - market
      - id
      - translation_status
      title: TranslationResponse
      description: Schema for translation response.
    TranslationUpdate:
      properties:
        name_en:
          anyOf:
          - type: string
          - type: 'null'
          title: Name En
        name_zh_tw:
          anyOf:
          - type: string
          - type: 'null'
          title: Name Zh Tw
        translation_status:
          anyOf:
          - type: string
            enum:
            - pending
            - approved
            - auto
          - type: 'null'
          title: Translation Status
      type: object
      title: TranslationUpdate
      description: Schema for updating an existing translation.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer